+++ src/runtime.rs
@@ -3,11 +3,17 @@ use crate::Runtime;
 /// Initialize the OCaml runtime, the runtime will be
 /// freed when the value goes out of scope
 pub fn init() -> Runtime {
+    #[cfg(not(feature = "no-std"))]
+    crate::initial_setup();
+
     Runtime::init()
 }
 
 /// Initialize the OCaml runtime
 pub fn init_persistent() {
+    #[cfg(not(feature = "no-std"))]
+    crate::initial_setup();
+
     Runtime::init_persistent()
 }
 
@@ -31,4 +37,10 @@ pub unsafe fn gc_full_major() {
 /// Run compaction
 pub unsafe fn gc_compact() {
     ocaml_sys::caml_gc_compaction(ocaml_sys::UNIT);
+}
+
+/// Initialize the OCaml runtime when
+pub fn startup() -> Runtime {
+    crate::macros::initial_setup();
+    Runtime::init()
 }
